requirements.txt
fastapi==0.104.1
httpx==0.25.2
pydantic==2.5.0
google-generativeai